Title: 소수의 거듭제곱, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240729/소수의 거듭제곱/power-of-prime-number.java
@@ -28,7 +28,7 @@ public class Main {
                 continue;
             }
 
-            int j = i;
+            long j = i;
             while (j < b) {
                 j *= i;
                 if (j >= a && j <= b) {
